pretty large refactor
