+++ tomography_preprocessing/tilt_series_alignment/_generate_matrices.py
@@ -25,7 +25,7 @@ def generate_relion_matrices(
     tilt_series_metadata = utils.star.iterate_tilt_series_metadata(
         tilt_series_star_file=tilt_series_star_file,
     )
-    image_directory, alignments_directory = \
+    tilt_series_directory, external_directory, alignments_directory = \
         create_alignment_job_directory_structure(output_directory)
     tilt_image_star_files = []
 
@@ -40,7 +40,7 @@ def generate_relion_matrices(
                 tilt_image_df['rlnMicrographName'][0])[:2],
             tomogram_dimensions=np.array(tomogram_dimensions)
         )
-        tilt_image_star_file = image_directory / f'{tilt_series_id}.star'
+        tilt_image_star_file = alignments_directory / f'{tilt_series_id}.star'
         tilt_image_star_files.append(tilt_image_star_file)
         projection_matrix_labels = [f'rlnTomoProj{ax}' for ax in 'XYZW']
         for idx, label in enumerate(projection_matrix_labels):
