+++ tomography_preprocessing/tilt_series_alignment/imod/_utils.py
@@ -1,12 +1,9 @@
 import os
 from functools import lru_cache
 from pathlib import Path
+from typing import Tuple
 
 import numpy as np
-import pandas as pd
-import starfile
-
-from ... import utils
 
 
 def read_xf(file: os.PathLike) -> np.ndarray:
@@ -90,76 +87,45 @@ def calculate_specimen_shifts(xf: np.ndarray) -> np.ndarray:
 
 
 @lru_cache(maxsize=100)
-def get_etomo_basename(imod_directory: Path) -> str:
+def get_tilt_series_id(alignment_directory: Path) -> str:
     """Get the tilt-series stack basename from an Etomo directory."""
-    edf_files = list(imod_directory.glob('*.edf'))
+    edf_files = list(alignment_directory.glob('*.edf'))
     if len(edf_files) != 1:
         raise RuntimeError('Multiple Etomo directive files found')
     return edf_files[0].stem
 
 
-def get_xf_file(imod_directory: Path) -> Path:
-    """Get the xf file containing image transforms from an IMOD directory."""
-    return imod_directory / f'{get_etomo_basename(imod_directory)}.xf'
-
-
-def get_tlt_file(imod_directory: Path) -> Path:
-    """Get the tlt file containing tilt angles from an IMOD directory."""
-    return imod_directory / f'{get_etomo_basename(imod_directory)}.tlt'
-
-
-def get_edf_file(imod_directory: Path) -> Path:
+def get_edf_file(imod_directory: Path, tilt_series_id: str) -> Path:
     """Get the Etomo directive file from an IMOD directory."""
-    return imod_directory / f'{get_etomo_basename(imod_directory)}.edf'
+    return imod_directory / f'{tilt_series_id}.edf'
 
 
-def get_xyz_extrinsic_euler_angles(xf: np.ndarray, tilt_angles: np.ndarray) -> np.ndarray:
+def get_xf_file(alignment_directory: Path, tilt_series_id: str) -> Path:
+    """Get the xf file containing image transforms from an IMOD directory."""
+    return alignment_directory / f'{tilt_series_id}.xf'
+
+
+def get_tlt_file(alignment_directory: Path, tilt_series_id: str) -> Path:
+    """Get the tlt file containing tilt angles from an IMOD directory."""
+    return alignment_directory / f'{tilt_series_id}.tlt'
+
+
+def get_tilt_series_alignment_data(
+        alignment_directory: Path, tilt_series_id: str
+) -> Tuple[np.ndarray, np.ndarray]:
+    """Get IMOD tilt-series alignment data."""
+    xf = read_xf(get_xf_file(alignment_directory, tilt_series_id))
+    tlt = read_tlt(get_tlt_file(alignment_directory, tilt_series_id))
+    return xf, tlt
+
+
+def get_xyz_extrinsic_euler_angles(
+        imod_xf_data: np.ndarray, imod_tlt_data: np.ndarray
+) -> np.ndarray:
     """Get XYZ extrinsic Euler angles which rotate the specimen."""
-    euler_angles = np.zeros(shape=(len(xf), 3))
-    euler_angles[:, 1] = tilt_angles
-    euler_angles[:, 2] = get_in_plane_rotations(xf)
+    euler_angles = np.zeros(shape=(len(imod_xf_data), 3))
+    euler_angles[:, 1] = imod_tlt_data
+    euler_angles[:, 2] = get_in_plane_rotations(imod_xf_data)
     return euler_angles
 
 
-def write_relion_tilt_series_alignment_output(
-        tilt_image_df: pd.DataFrame,
-        tilt_series_id: str,
-        pixel_size: float,
-        imod_directory: Path,
-        output_star_file: Path,
-):
-    """Write output from a tilt-series alignment experiment."""
-    xf = read_xf(get_xf_file(imod_directory))
-    tlt = read_tlt(get_tlt_file(imod_directory))
-    shifts_px = calculate_specimen_shifts(xf)
-
-    tilt_image_df[['rlnTomoXShiftAngst', 'rlnTomoYShiftAngst']] = shifts_px * pixel_size
-    tilt_image_df[['rlnTomoXTilt', 'rlnTomoYTilt', 'rlnTomoZRot']] = \
-        get_xyz_extrinsic_euler_angles(xf, tlt)
-
-    starfile.write({tilt_series_id: tilt_image_df}, output_star_file)
-
-
-def write_aligned_tilt_series_star_file(
-        original_tilt_series_star_file: Path,
-        output_directory: Path,
-):
-    df = starfile.read(original_tilt_series_star_file, always_dict=True)['global']
-    tilt_series_metadata = list(
-        utils.star.iterate_tilt_series_metadata(original_tilt_series_star_file)
-    )
-    # update individual tilt series star files
-    df['rlnTomoTiltSeriesStarFile'] = [
-        output_directory / 'tilt_series' / f'{tilt_series_id}.star'
-        for tilt_series_id, _, _ in tilt_series_metadata
-    ]
-    df['EtomoDirectiveFile'] = [
-        output_directory / 'alignments' / tilt_series_id / f'{tilt_series_id}.edf'
-        for tilt_series_id, _, _ in tilt_series_metadata
-    ]
-    etomo_dir_exist = any(df['EtomoDirectiveFile'].apply(lambda x: Path(x).exists()))
-    if etomo_dir_exist is False:
-        df = df.drop(columns=['EtomoDirectiveFile'])
-    # check which output files were succesfully generated, take only those
-    df = df[df['rlnTomoTiltSeriesStarFile'].apply(lambda x: x.exists())]
-    starfile.write({'global': df}, output_directory / 'aligned_tilt_series.star')
